[BT] Add Brawler 64 support
Native mode fail to enable and HID mode button doen't match the NSO one.
There is no mode in which the controller assume it's own identity.
As such, there is not a straight forward way to identify it beside assuming
that a Switch N64 controller that failed to go into native mode is the Brawler 64.